Move title bar to its own file
