+++ examples/api/src-tauri/capabilities/base.json
@@ -17,8 +17,7 @@
     "core:default",
     "fs:default",
     "core:window:allow-minimize",
-    "core:window:allow-maximize",
-    "core:window:allow-unmaximize",
+    "core:window:allow-toggle-maximize",
     "core:window:allow-close",
     "core:window:allow-start-dragging",
     "notification:default",
