+++ examples/api/src/App.svelte
@@ -1,4 +1,5 @@
 <script>
+  import { onMount, tick } from 'svelte'
   import { writable } from 'svelte/store'
   import { getCurrentWindow } from '@tauri-apps/api/window'
   import { getCurrentWebview } from '@tauri-apps/api/webview'
@@ -23,10 +24,9 @@
   import Biometric from './views/Biometric.svelte'
   import Geolocation from './views/Geolocation.svelte'
   import Haptics from './views/Haptics.svelte'
-
-  import { onMount, tick } from 'svelte'
-  import { ask } from '@tauri-apps/plugin-dialog'
   import Nfc from './views/Nfc.svelte'
+
+  import TitleBar from './lib/TitleBar.svelte'
 
   const appWindow = getCurrentWindow()
 
@@ -155,44 +155,9 @@
     selected = view
   }
 
-  // Window controls
-  let isWindowMaximized
-  onMount(async () => {
-    isWindowMaximized = await appWindow.isMaximized()
-    appWindow.onResized(async () => {
-      isWindowMaximized = await appWindow.isMaximized()
-    })
-  })
-
-  function minimize() {
-    appWindow.minimize()
-  }
-
-  async function toggleMaximize() {
-    ;(await appWindow.isMaximized())
-      ? appWindow.unmaximize()
-      : appWindow.maximize()
-  }
-
-  let confirmed_close = false
-  async function close() {
-    if (!confirmed_close) {
-      confirmed_close = await ask(
-        'Are you sure that you want to close this window?',
-        {
-          title: 'Tauri API'
-        }
-      )
-      if (confirmed_close) {
-        appWindow.close()
-      }
-    }
-  }
-
   // dark/light
-  let isDark
+  let isDark = localStorage.getItem('theme') == 'dark'
   onMount(() => {
-    isDark = localStorage && localStorage.getItem('theme') == 'dark'
     applyTheme(isDark)
   })
   function applyTheme(isDark) {
@@ -265,10 +230,7 @@
     document.addEventListener('mousemove', moveHandler)
   }
 
-  let isWindows
-  onMount(async () => {
-    isWindows = (await os.platform()) === 'windows'
-  })
+  let isWindows = os.platform() === 'windows'
 
   // mobile
   let isSideBarOpen = false
@@ -340,62 +302,7 @@
 
 <!-- custom titlebar for Windows -->
 {#if isWindows}
-  <div
-    class="w-screen select-none h-8 flex justify-between items-center absolute text-primaryText dark:text-darkPrimaryText"
-    data-tauri-drag-region
-  >
-    <span
-      class="h-100% pl-2 flex-1 flex items-center lt-sm:pl-10 lt-lg:text-darkPrimaryText [app-region:drag]"
-      >Tauri API Validation</span
-    >
-    <span
-      class="
-      h-100%
-      children:h-100% children:w-12 children:inline-flex
-      children:items-center children:justify-center"
-    >
-      <button
-        aria-label="Toggle dark mode"
-        title={isDark ? 'Switch to Light mode' : 'Switch to Dark mode'}
-        class="border-none hover:bg-hoverOverlay active:bg-hoverOverlayDarker dark:hover:bg-darkHoverOverlay dark:active:bg-darkHoverOverlayDarker"
-        on:click={toggleDark}
-      >
-        {#if isDark}
-          <div class="i-ph-sun"></div>
-        {:else}
-          <div class="i-ph-moon"></div>
-        {/if}
-      </button>
-      <button
-        aria-label="Minimize window"
-        title="Minimize"
-        class="border-none hover:bg-hoverOverlay active:bg-hoverOverlayDarker dark:hover:bg-darkHoverOverlay dark:active:bg-darkHoverOverlayDarker"
-        on:click={minimize}
-      >
-        <div class="i-codicon-chrome-minimize"></div>
-      </button>
-      <button
-        aria-label="Maximize window"
-        title={isWindowMaximized ? 'Restore' : 'Maximize'}
-        class="border-none hover:bg-hoverOverlay active:bg-hoverOverlayDarker dark:hover:bg-darkHoverOverlay dark:active:bg-darkHoverOverlayDarker"
-        on:click={toggleMaximize}
-      >
-        {#if isWindowMaximized}
-          <div class="i-codicon-chrome-restore"></div>
-        {:else}
-          <div class="i-codicon-chrome-maximize"></div>
-        {/if}
-      </button>
-      <button
-        aria-label="Close window"
-        title="Close"
-        class="border-none hover:bg-red-700 dark:hover:bg-red-700 hover:text-darkPrimaryText active:bg-red-700/90 dark:active:bg-red-700/90 active:text-darkPrimaryText"
-        on:click={close}
-      >
-        <div class="i-codicon-chrome-close"></div>
-      </button>
-    </span>
-  </div>
+  <TitleBar {appWindow} {isDark} {toggleDark} />
 {/if}
 
 <!-- Sidebar toggle, only visible on small screens -->
